Reinsert if reuse time is equal to the threshold
Summary:
We want to reinsert if reuse time is equal to the threshold. We define the threshold as some multiple of bucket width. For example, we have 15 minute buckets, and reuse threshold is 45 minutes. We used to reinsert keys in bucket 0 and 1, which is reuse time of 0-15 and 15-30 but not reinsert 30-45 because we approximate reuse time to be the upper limit of the bucket that the key is in.

Fix some tests.

Reviewed By: AlnisM

Differential Revision: D86172687

fbshipit-source-id: b45e89872e3fe0de457fd213fa2c37aa27e2f278