Make ref type arg as optional for Index::persist() and Index::recover()
Summary:
(Reattempt for D81760167.
It was urgently reverted together with D81619566 for S576441)

With the later diff(s) to implement persist/recover using shared memory for FixedSizeIndex, it doesn't always need RecordWriter, RecordReader arg for persist/recover.
But we don't want to eliminate it from the interface since current index structure (SparseMapIndex) still need it.
So I just made it as optional reference argument.

Reviewed By: rlyerly

Differential Revision: D85916189

fbshipit-source-id: 2af445648267ee2152893c0f8292bdad91988c88