Adding NavyPersistParams for the option to use shm for navy persistence
Summary:
Creating NavyPersistParams for the option on how to persist (using flash or Shm).
SparseMapIndex will only support flash persistence while FixedSizeIndex will support Shm. (FixedSizeIndex is planned to support flash persistence too in the near future).
It's not put directly into navy config, since it's not directly user facing config.

This diff is just defining NavyPersistParams and passing it down to navy and BC, but it's not used yet.

More context is as below:

(D81619566 had to be reverted due to S576441.
Looked into it and found it's failing to create a folder for separate ShmManager for navy.
Though it worked for ucache, it didn't work for other use case and creating a separate folder and ShmManager within our navy code seems to be bad idea.

This diff will create NavyPersistParams and hand it down to navy (and eventually BC) so that navy and BC just can use the given ShmManager.
That way, If we want to create a separate ShmManager, we can create one and pass it down. If we want to share existing one (that we are using for DRAM cache), we can also do it. )

Reviewed By: rlyerly

Differential Revision: D85915614

fbshipit-source-id: e21a9eeb68ece8cf4904f8eb2e2bd570e64b7dfe