prevent memory monitor from advising away more than maxLimitPercent
Summary:
Currently, it is possible for MemoryMonitor to advise away more than maxLimitPercent_ of the cache. MemoryMonitor just checks if the amount of currently advised away memory is higher than maxLimitPercent_. It does not take into account how much memory will be advised away in total after performing the slab releases.

If MemoryMonitor is configured in a way where maxLimitPercent_ is set to 80% and in each iteration 10% of the memory is advised away, then after each 8 iterations, we will be at exactly 80%, but MemoryMonitor will still do one more iteration, resulting in 90% of the memory being advised away.

I change this behavior such that when making the decision if and how much to advise away we also take into account how much is going to be advised away. This ensures that we never advise away more than maxLimitPercent_ of the cache.

Reviewed By: pbhandar2, rlyerly

Differential Revision: D87043177

fbshipit-source-id: d05becebd76fa094667249306d6410290ca16a82