Add latency numbers for BlockCache APIs
Summary: We have stats on the time it takes to issue IO operations, but we don't have stats on the time it takes to call the actual BlockCache APIs (which usually have a layer of buffering before the IO op).  Add latency stats for insert, lookup & remove APIs & print them in cachebench.

Reviewed By: AlnisM

Differential Revision: D89493277

fbshipit-source-id: b2ba54170c6d0859edd8fe52ecf61e1000ad897e